refactor(v3): move issue_branch_resolver from utils to services
Move issue_branch_resolver.py from vibe3.utils to vibe3.services
to align with its role as a business logic service.

Changes:
- Move src/vibe3/utils/issue_branch_resolver.py to services/
- Update 5 source files with new import path
- Update test imports to use services path
- Fix import sorting in affected files

No logic changes, pure file move.

Ref: #1266
