+++ src/vibe3/utils/pr_branch_resolver.py
@@ -7,7 +7,7 @@ import typer
 from vibe3.clients.github_client import GitHubClient
 from vibe3.exceptions import UserError
 from vibe3.services.flow_service import FlowService
-from vibe3.utils.issue_branch_resolver import resolve_issue_branch_input
+from vibe3.services.issue_branch_resolver import resolve_issue_branch_input
 
 
 def resolve_branch_from_pr(
